requirements.txt
amqp==1.4.9
aniso8601==7.0.0
anyjson==0.3.3
appdirs==1.4.4
argcomplete==1.12.2
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
curlylint==0.12.2
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-celery==3.3.1
django-compressor==2.4.1
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-google-tag-manager==0.0.5
django-livereload-server==0.3.2
django-livesync==0.5
django-maintenance-mode==0.16.0
django-rename-app==0.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
escpos==1.8
ffmpeg==1.4
Flask==1.1.2
future==0.18.2
geoip2==2.9.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
huepy==1.2.1
hurry==1.1
hurry.filesize==0.9
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==3.0.37
libusb==1.0.23b7
MarkupSafe==1.1.1
maxminddb==2.0.3
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
parsy==1.1.0
pathspec==0.8.1
pep8==1.7.1
Pillow==8.1.2
pluggy==0.13.1
proglog==0.1.9
promise==2.3
psycopg2==2.8.6
py==1.10.0
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
PyESCPOS==0.3
pyflakes==2.2.0
PyGetWindow==0.0.9
PyJWT==2.0.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyserial==3.5
PySocks==1.7.1
pytest==6.2.3
python-barcode==0.13.1
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2021.1
pyusb==1.1.1
pywin32==301
PyYAML==5.4.1
qrcode==6.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.13.2
rjsmin==1.1.0
Rx==1.6.1
schedule==1.1.0
singledispatch==3.7.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.60.0
Unidecode==1.2.0
urllib3==1.26.4
uuid==1.30
viivakoodi==0.8.0
watchdog==2.0.2
websocket-client==0.58.0
Werkzeug==1.0.1
wfastcgi==3.0.0
whitenoise==5.2.0
